Запустил retrofit2, добавил функцию receiptToDb исправил тип для receipt.fiscalSign с float на double для правильной генерации ключа idFns

+++ app/src/main/java/com/vk/vsvans/BlogShop/model/fns/Receipt.kt
@@ -10,7 +10,7 @@ data class Receipt(
     val fiscalDocumentFormatVer: Int,
     val fiscalDocumentNumber: Int,
     val fiscalDriveNumber: String,
-    val fiscalSign: Float,
+    val fiscalSign: Double,
     val fnsUrl: String,
     val items: List<Item>,
     val kktRegId: String,
